Changed the size of the address in binaries and modified the argument number of Mult operation to handle div/modulo correctly

+++ includes/transformation/irGenerator.h
@@ -11,7 +11,7 @@
 #include <types.h>
 
 unsigned int irGenerator(DBTPlateform *platform,
-		uint16 addressInBinaries,
+		uint32 addressInBinaries,
 		uint32 blockSize,
 		uint32 globalVariableCounter);
